Revert "Fix cross command in README"
Command works fine when CrossPerProjectPlugin is enabled which is documented
in the appropriate section. This method is also probably preferrable to using
sbt-doge's 'so' anyways since it plays nicer with other plugins, e.g.
sbt-rig https://github.com/Verizon/sbt-rig which relies on '++' and should
reasonably not depend on sbt-doge.

This reverts commit f5dfba0510dc721d7ac756fb0ac858ff7cea0cc6.